Basic enemy pathing ai done (VERY BASIC)
Done the basic enemy ai for now. It goes towards the player using A*, but a bunch of things aren't quite implemented yet.
eg. X and Y dimention, but in unity that means up/down and left/right as Z is forward/back?????
shooting probably isn't working
I can add a thing to make the enemy stop a distance away from the player