Added line of sight pathing
Enemy AI will now only go to player if it is in line of sight (LOS). If LOS is blocked by an obstacle, the enemy will not go after the player. If the player was in LOS, but then is blocked by an obstacle it will go to the player's last seen location.